@@ -18,11 +18,11 @@ import Alamofire
   You can create instances of this provider with a custom
   [Alamofire.Manager](http://cocoadocs.org/docsets/Alamofire/1.3.0/Classes/Manager.html)
   in order to control caching, certificate validation rules, etc. For example, here is a `Service` that will
-  not use the cell network:
+  use a URLCache and will not use the cell network:
 
       class MyAPI: Service {
           init() {
-              let configuration = URLSessionConfiguration.ephemeral
+              let configuration = URLSessionConfiguration.defaultSessionConfiguration()
               configuration.allowsCellularAccess = false
               super.init(
                   baseURL: "http://foo.bar/v1",
@@ -32,62 +32,50 @@ import Alamofire
 */
 public struct AlamofireProvider: NetworkingProvider
     {
-    public let session: Alamofire.Session
+    public let manager: Alamofire.SessionManager
 
-    public init(session: Alamofire.Session = Session.default)
-        { self.session = session }
+    public init(manager: Alamofire.SessionManager = SessionManager.default)
+        { self.manager = manager }
 
     public init(configuration: URLSessionConfiguration)
-        { self.init(session: Alamofire.Session(configuration: configuration)) }
+        { self.init(manager: Alamofire.SessionManager(configuration: configuration)) }
 
     public func startRequest(
             _ request: URLRequest,
             completion: @escaping RequestNetworkingCompletionCallback)
         -> RequestNetworking
         {
-        AlamofireRequestNetworking(
-            session: session,
-            request: session.request(request)
+        return AlamofireRequestNetworking(
+            manager.request(request)
                 .response { completion($0.response, $0.data, $0.error) })
         }
     }
 
 internal struct AlamofireRequestNetworking: RequestNetworking, SessionTaskContainer
     {
-    let session: Alamofire.Session
-    let alamofireRequest: Alamofire.Request
+    internal var alamofireRequest: Alamofire.Request
 
-    init(session: Alamofire.Session, request alamofireRequest: Alamofire.Request)
+    init(_ alamofireRequest: Alamofire.Request)
         {
-        self.session = session
         self.alamofireRequest = alamofireRequest
+        if let requestTask = alamofireRequest.task, case .suspended = requestTask.state
+            {
+            alamofireRequest.resume()   // in case manager.startRequestsImmediately is false
+            }
         }
 
     var task: URLSessionTask
         {
-        session.rootQueue.sync
-            {
-            guard let requestTask = alamofireRequest.task else
-                { return ZeroProgressURLSessionTask() }
-            if requestTask.state == .suspended
-                { alamofireRequest.resume() }   // in case session.startRequestsImmediately is false
-            return requestTask
-            }
+        return alamofireRequest.task!
         }
 
     func cancel()
         { alamofireRequest.cancel() }
     }
 
-extension Alamofire.Session: NetworkingProviderConvertible
+extension Alamofire.SessionManager: NetworkingProviderConvertible
     {
     /// You can pass an `AlamoFire.Manager` when creating a `Service`.
     public var siestaNetworkingProvider: NetworkingProvider
-        { return AlamofireProvider(session: self) }
-    }
-
-private class ZeroProgressURLSessionTask: URLSessionTask
-    {
-    override var countOfBytesSent: Int64
-        { return 0 }
+        { return AlamofireProvider(manager: self) }
     }
