@@ -9,7 +9,6 @@
 #import <Siesta/Siesta.h>
 #import <Quick/Quick.h>
 #import <Nimble/Nimble.h>
-#import <Nocilla/Nocilla.h>
 #import "SiestaTests-Swift.h"
 
 @interface ObjcCompatibilitySpec : QuickSpec
@@ -26,10 +25,12 @@
     // These specs are mostly just concerned with whether the code compiles.
     // They don't make many assertions about behavior, which is tested elsewhere.
 
-    __block BOSService *service;
+    __block TestService *service;
     __block BOSResource *resource;
 
-    __block id _;  // Fake Swift’s `_ = foo()` idiom for @discardableResult functions
+    __block id _;  // Fake version of Swift’s `_ = foo()` idiom for non-@discardableResult functions
+
+    __block NSMutableArray *allRequests;
 
     beforeEach(^
         {
@@ -37,50 +38,92 @@
         resource = [service resource:@"/foo"];
         });
 
+    beforeEach(^
+        {
+        allRequests = [NSMutableArray array];
+        });
+
     afterEach(^
         {
+        [service awaitAllRequests];  // because unlike the Swift specs, the individual specs here don't wait for them
+        [NetworkStub clearAll];
         service = nil;
         resource = nil;
         });
-
-    beforeSuite(^{ [LSNocilla.sharedInstance start]; });
-    afterSuite( ^{ [LSNocilla.sharedInstance stop]; });
-    afterEach(  ^{ [LSNocilla.sharedInstance clearStubs]; });
 
     it(@"handles resource paths", ^
         {
         _ = [resource child:@"bar"];
         _ = [resource relative:@"../bar"];
         _ = [resource withParam:@"foo" value:@"bar"];
+        _ = [resource withParams:@{@"foo": @"bar", @"baz": NSNull.null}];
         });
 
-    it(@"handles requests", ^
+    it(@"handles basic request", ^
         {
-        stubRequest(@"GET", @"http://example.api/foo").andReturn(200);
-        stubRequest(@"POST", @"http://example.api/foo").andReturn(200);
-
+        [NetworkStub addForMethod:@"GET" resource:resource returningStatusCode:200];
         expect([resource loadIfNeeded]).notTo(beNil());
         _ = [resource load];
+        });
+
+    it(@"handles request mutation", ^
+        {
+        [NetworkStub addForMethod:@"POST" resource:resource returningStatusCode:200];
         _ = [resource requestWithMethod:@"DELETE" data:[[NSData alloc] init] contentType:@"foo/bar" requestMutation:
             ^(NSMutableURLRequest *req)
                 { req.HTTPMethod = @"POST"; }];
-        _ = [resource requestWithMethod:@"POST" json:@{@"foo": @"bar"}];
-        _ = [resource requestWithMethod:@"POST" json:@{@"foo": @"bar"} contentType:@"foo/bar" requestMutation:nil];
-        _ = [resource requestWithMethod:@"POST" text:@"Ahoy"];
-        _ = [resource requestWithMethod:@"POST" text:@"Ahoy" contentType:@"foo/bar" encoding:NSASCIIStringEncoding requestMutation:nil];
-        _ = [resource requestWithMethod:@"POST" urlEncoded:@{@"foo": @"bar"} requestMutation:nil];
-        _ = [resource loadUsingRequest:[resource requestWithMethod:@"POST" json:@{@"foo": @"bar"}]];
+        });
 
-        XCTestExpectation *expectation = [[QuickSpec current] expectationWithDescription:@"network calls finished"];
+    it(@"handles JSON request", ^
+        {
+        [NetworkStub addForMethod:@"POST" resource:resource headers:@{@"Content-Type": @"application/json"} body:@"{\"foo\":\"bar\"}" returningStatusCode:200];
+        _ = [resource requestWithMethod:@"POST" json:@{@"foo": @"bar"}];
+        });
+
+    it(@"handles JSON request with custom content type", ^
+        {
+        [NetworkStub addForMethod:@"POST" resource:resource headers:@{@"Content-Type": @"foo/bar"} body:@"{\"foo\":\"bar\"}" returningStatusCode:200];
+        _ = [resource requestWithMethod:@"POST" json:@{@"foo": @"bar"} contentType:@"foo/bar" requestMutation:nil];
+        });
+
+    it(@"handles text request", ^
+        {
+        [NetworkStub addForMethod:@"POST" resource:resource headers:@{@"Content-Type": @"text/plain; charset=utf-8"} body:@"Ahoy" returningStatusCode:200];
+        _ = [resource requestWithMethod:@"POST" text:@"Ahoy"];
+        });
+
+    it(@"handles test request with custom content type and encoding", ^
+        {
+        [NetworkStub addForMethod:@"POST" resource:resource headers:@{@"Content-Type": @"foo/bar; charset=us-ascii"} body:@"Ahoy" returningStatusCode:200];
+        _ = [resource requestWithMethod:@"POST" text:@"Ahoy" contentType:@"foo/bar" encoding:NSASCIIStringEncoding requestMutation:nil];
+        });
+
+    it(@"handles URL-encoded request", ^
+        {
+        [NetworkStub addForMethod:@"POST" resource:resource headers:@{@"Content-Type": @"application/x-www-form-urlencoded"} body:@"foo=bar" returningStatusCode:200];
+        _ = [resource requestWithMethod:@"POST" urlEncoded:@{@"foo": @"bar"} requestMutation:nil];
+        });
+
+    it(@"handles loadUsingRequest:", ^
+        {
+        [NetworkStub addForMethod:@"POST" resource:resource headers:@{} body:@"{\"foo\":\"bar\"}" returningStatusCode:200];
+        _ = [resource loadUsingRequest:[resource requestWithMethod:@"POST" json:@{@"foo": @"bar"}]];
+        });
+
+    it(@"handles callbacks", ^
+        {
+        [NetworkStub addForMethod:@"GET" resource:resource returningStatusCode:200];
+        XCTestExpectation *expectation = [[QuickSpec current] expectationWithDescription:@"callback called"];
         BOSRequest *req = [[[[[[resource load]
             onCompletion:
                 ^(BOSEntity *entity, BOSError *error)
                     { [expectation fulfill]; }]
             onSuccess: ^(BOSEntity *entity) { }]
-            onNewData: ^(BOSEntity *entity) { }]  // TODO: This line leaks Resource instances, but the previous line doesn't. ‽‽‽
+            onNewData: ^(BOSEntity *entity) { }]
             onNotModified: ^{ }]
             onFailure: ^(BOSError *error) { }];
         [[QuickSpec current] waitForExpectationsWithTimeout:1 handler:nil];
+
         [req cancel];
         });
 
@@ -114,10 +157,7 @@
 
     it(@"handles resource data", ^
         {
-        stubRequest(@"GET", @"http://example.api/foo")
-            .andReturn(200)
-            .withHeader(@"Content-type", @"application/json")
-            .withBody(@"{\"foo\": \"bar\"}");
+        [NetworkStub addForMethod:@"GET" resource:resource returningStatusCode:200 headers:@{@"Content-type": @"application/json"} body:@"{\"foo\": \"bar\"}"];
 
         XCTestExpectation *expectation = [[QuickSpec current] expectationWithDescription:@"network calls finished"];
         _ = [[resource load] onSuccess:^(BOSEntity *entity) { [expectation fulfill]; }];
@@ -143,7 +183,7 @@
 
     it(@"handles HTTP errors", ^
         {
-        stubRequest(@"GET", @"http://example.api/foo").andReturn(507);
+        [NetworkStub addForMethod:@"GET" resource:resource returningStatusCode:507];
 
         XCTestExpectation *expectation = [[QuickSpec current] expectationWithDescription:@"network calls finished"];
         _ = [[resource load] onFailure:^(BOSError *error) { [expectation fulfill]; }];
@@ -185,10 +225,7 @@
             blockObserverCalls++;
         }];
 
-        stubRequest(@"GET", @"http://example.api/foo")
-            .andReturn(200)
-            .withHeader(@"Content-type", @"application/json")
-            .withBody(@"{\"foo\": \"bar\"}");
+        [NetworkStub addForMethod:@"GET" resource:resource returningStatusCode:200 headers:@{@"Content-type": @"application/json"} body:@"{\"foo\": \"bar\"}"];
 
         XCTestExpectation *expectation = [[QuickSpec current] expectationWithDescription:@"network calls finished"];
         _ = [[resource load] onSuccess:^(BOSEntity *entity) { [expectation fulfill]; }];
